SportsClassification/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cv_projects"
version = "0.1.0"
description = "Computer vision projects"
authors = [
    { name = "Michał Bogusławski", email = "boguslawski.m.j@gmail.com" }
]
requires-python = ">=3.8"
readme = "README.md"
keywords = ["computer vision", "sports", "classification", "detection"]
license = { file = "LICENSE" }

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy",
    "Pillow",
    "PyYAML",
    "tqdm",
    "torcheval",
    "pandas",
    "kagglehub",
    "pytest"
]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*"]
